restore dropped LVD objects and update diameters; many more updates on ssl model; utilities moved around
