+++ py/SGA/data/SGA2025/SGA2025-crossid-errors.csv
@@ -1355,7 +1355,7 @@ NGC 3303,31508,NGC 3303 NED02,93104,5.103,GPair
 NGC 3303 NED02,93104,NGC 3303 NED01,-99,11.231,G
 UGC 06865,37282,UGC 06865 NED02,-99,18.766,GPair
 VV 185,37892,VV 185a,-99,4.848,GPair
-UGC 07230,39014,2MASX J12133763+1607095,-99,19.237,GPair
+UGC 07230,39014,SDSS J121337.65+160710.2,-99,19.237,GPair
 UGC 07891,42760,UGC 07891 NED01,-99,11.133,GPair
 IC 3862,43844,SDSS J125353.12+360512.7,-99,3.808,GPair
 IC 4271,47334,IC 4271 NED01,-99,8.289,GPair
