+++ py/SGA/data/SGA2025/SGA2025-vi-actions.csv
@@ -1,11 +1,5 @@
 # Object-level actions based on VI (visual inspections).
 objname,action,reason,comment
-# LVD dwarfs with no structural information; see
-A0952+69,drop,extreme lsb,From LVD sample.
-Clump III,drop,extreme lsb,From LVD sample.
-Clump I,drop,extreme lsb,From LVD sample.
-KK 35,drop,extreme lsb,From LVD sample.
-KK 208,drop,extreme lsb,From LVD sample.
 #
 SDSS J125839.45+273922.5,drop,faint,In '1256+27W01' group.
 [APS2006] J194.676010+27.66179,drop,faint,In '1256+27W01' group.
@@ -910,6 +904,7 @@ SDSS J002041.45+083701.2,drop,shred,Shred or part of 'JKB129'.
 GALEXASC J131520.11+473234.3,drop,shred,Shred or part of 'DDO 169NW'.
 SDSS J131704.45+375702.5,drop,shred,Shred or part of 'CGCG 189-050'.
 2MASX J14432661-4442199,drop,shred,Shred or part of 'ESO272-025'.
+SDSS J161349.24+542205.5,drop,shred,Shred or part of 'PGC1 2801026 NED001'.
 #
 UGC 08691 NED02,drop,star,Foreground star in 'UGC 08691 NED01'.
 UGC 08691 NED03,drop,star,Foreground star in 'UGC 08691 NED01'.
