+++ py/SGA/data/SGA2025/SGA2025-vi-properties.csv
@@ -234,10 +234,16 @@ ESO-LV 1570281,66.559594,-53.151251,-99.,-99.,-99.,-99.,No,Moved to the LS/DR9 p
 ESO-LV 4170091,44.7971,-32.0752,-99.,-99.,-99.,-99.,No,Moved to the LS/DR9 position.
 ESO-LV 2360021,-99.,-99.,0.2,0.,1.,-99.,No,Estimated diameter.
 SDSS J121337.65+160710.2,-99.,-99.,0.7,0.,1.,0.7,No,Estimated diameter.
-PGC 725719,200.73749999999998,-29.739333333333335,-99.,-99.,-99.,-99.,No,LVD position shifted to NED coordinates.
-KDG 73,163.23779166666665,69.54938888888888,-99.,-99.,-99.,-99.,No,LVD position shifted to NED coordinates.
-JKB142,26.45083,16.37806,-99.,-99.,-99.,-99.,No,LVD position shifted to NED coordinates.
 DDO 169NW,198.833750,47.543611,-99.,-99.,-99.,-99.,No,LVD position shifted to the coordinates in 2013AJ....145..101K.
-CGCG 189-050,199.26954166666664,37.95291666666667,-99.,-99.,-99.,-99.,No,LVD position shifted to NED coordinates.
-KKH 57,150.06416666666667,63.18372222222222,-99.,-99.,-99.,-99.,No,LVD position shifted to NED coordinates.
-NGC 5194,202.46962499999995,47.195166666666665,-99.,-99.,-99.,-99.,No,LVD position shifted to NED coordinates.
+KDG 73,163.23779166666665,69.54938888888888,-99.,-99.,-99.,-99.,No,LVD dwarf moved to NED position.
+Clump I,-99.,-99.,0.18,0.,1.,-99.,No,LVD dwarf with geometry taken from the LVG database.
+Clump III,-99.,-99.,0.10,0.,1.,-99.,No,LVD dwarf with geometry taken from the LVG database.
+A0952+69,-99.,-99.,1.82,0.,1.,-99.,No,LVD dwarf with geometry taken from the LVG database.
+KK 35,-99.,-99.,2.51,0.,1.,-99.,No,LVD dwarf with geometry taken from the LVG database.
+KK 208,-99.,-99.,5.89,0.,1.,-99.,No,LVD dwarf with geometry taken from the LVG database.
+DDO 147,-99.,-99.,-99.,74.,-99.,-99.,No,LVD dwarf with updated position angle.
+NGC 4190,-99.,-99.,-99.,30.,-99.,-99.,No,LVD dwarf with updated position angle.
+DDO126,-99.,-99.,-99.,145.,-99.,-99.,No,LVD dwarf with updated position angle.
+DDO 133,-99.,-99.,-99.,160.,-99.,-99.,No,LVD dwarf with updated position angle.
+DDO 165,-99.,-99.,-99.,90.,-99.,-99.,No,LVD dwarf with updated position angle.
+Holmberg I,-99.,-99.,-99.,105.,-99.,-99.,No,LVD dwarf with updated position angle.
